refactor: `ServiceDetail`에서 `ServiceAction` 분리
- 토스트 위치가 이상한데 YDS 구버전이 남아서 그런건지 확인 필요
